Added old, original application code.
This is a little backwards, because the master
branch code actually came from this, previous to
my git days. I am currently doing some serious
house cleaning and found that I still have a
project running on this code. So I need to store
it until I can migrate to the new code in master.